B41-F24-PA34

+++ backend/src/main/groovy/com/backend/gbp/domain/projects/ProjectUpdatesMaterials.groovy
@@ -49,15 +49,15 @@ class ProjectUpdatesMaterials extends AbstractAuditingEntity implements Serializ
 
 	@GraphQLQuery
 	@Column(name = "on_hand")
-	Integer onHand
+	BigDecimal onHand
 
 	@GraphQLQuery
 	@Column(name = "qty")
-	Integer qty
+	BigDecimal qty
 
 	@GraphQLQuery
 	@Column(name = "balance")
-	Integer balance
+	BigDecimal balance
 
 	@GraphQLQuery
 	@Column(name = "w_cost")
